v 0.7

+++ app/src/main/java/com/qoiu/dailytaskmotivator/domain/TaskInteractor.kt
@@ -1,40 +1,51 @@
 package com.qoiu.dailytaskmotivator.domain
 
-import com.qoiu.dailytaskmotivator.data.TaskDb
-import com.qoiu.dailytaskmotivator.data.TaskRepository
+import com.qoiu.dailytaskmotivator.ResourceProvider
+import com.qoiu.dailytaskmotivator.domain.entities.DefaultTasks
+import com.qoiu.dailytaskmotivator.domain.entities.Task
 
 interface TaskInteractor {
-    suspend fun loadTask(): List<TaskDb>
-    suspend fun removeTask(task: TaskDb)
-    suspend fun update(task: TaskDb)
+    suspend fun loadTask(): List<Task>
+    suspend fun removeTask(task: Task)
+    suspend fun save(data: Task)
 
-    class Base(private val repository: TaskRepository) : TaskInteractor {
-        override suspend fun loadTask(): List<TaskDb> {
+    class Base(private val repository: Repository<Task>, private val stringProvider: ResourceProvider.StringProvider) : TaskInteractor {
+        override suspend fun loadTask(): List<Task> {
             val tasks = repository.fetchData()
-            val list = mutableListOf<TaskDb>()
+            if(tasks.isEmpty()){
+                DefaultTasks(stringProvider).getDefault().forEach { save(it) }
+                return DefaultTasks(stringProvider).getDefault()
+            }
+            val list = mutableListOf<Task>()
             tasks.forEach {
                 if (it.deadline > 0)
                     TaskGoldCoficient.Base(it).modify()
                 if (it.title == "" || it.reward == 0) {
                     removeTask(it)
-                } else if (!(it.dailyTask && it.expiredAt <= TaskCalendar().today().time)) {
+                } else if (!(it.dailyTask && it.expired <= TaskCalendar().today().time)) {
                     list.add(it)
                 }
+            }
+            list.sortBy {
+                it.category
             }
             return list
         }
 
-        override suspend fun removeTask(task: TaskDb) {
-            if (!task.dailyTask) {
-                repository.remove(task)
-            } else {
-                task.expiredAt = TaskCalendar().today().time
-                update(task)
-            }
+        override suspend fun save(data: Task) {
+            repository.save(data)
+            loadTask()
         }
 
-        override suspend fun update(task: TaskDb) {
-            repository.update(task)
+        override suspend fun removeTask(task: Task) {
+            if (!task.dailyTask && !task.reusable) {
+                repository.remove(task)
+            } else {
+                if(!task.reusable) {
+                    task.expired = TaskCalendar().today().time
+                    save(task)
+                }
+            }
         }
     }
 }
